+++ aldrin/CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Rename the types `Reply` and `low_level::Reply` to `PendingReply`.
+- `low_level::PendingReply` now resolves to `low_level::Reply` instead of directly to the call's
+  result.
 
 ## [0.11.0] - 2025-01-07